% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate adjusted R-squared between predicted and true values}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A vector of predicted values}

\item{truth}{A vector of true values}

\item{n_p}{The number of model parameters, excluding the intercept}
}
\value{
The adjusted R-squared statistic between \code{pred} and \code{truth}
}
\description{
This function calculates the adjusted R-squared statistic between a vector of predicted values \code{pred} and a vector of true values \code{truth}, given the number of model parameters \code{n_p}.
}
\examples{
adj_R2(c(1,2,3,4), c(1,3,2,5), 1)

}
