% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{bmi510final}
\alias{bmi510final}
\title{Final Project Package for BMI 510}
\description{
This package contains the functions required to complete the final project of BMI 510.
}
\author{
Elijah Chou \email{elijah.chou@emory.edu}
}
