% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Get the classes of variables in a tibble}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble to get the variable classes from}
}
\value{
A character vector containing the classes of each variable in the tibble
}
\description{
This function takes a tibble \code{x} and returns a character vector containing the classes of each variable in the tibble.
}
\examples{
classes(mtcars)

}
