% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Scale the numeric variables in a tibble}
\usage{
df_scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A tibble to scale}

\item{center}{Logical. Should the variables be centered? (default = TRUE)}

\item{scale}{Logical. Should the variables be scaled to have unit variance? (default = TRUE)}
}
\value{
A tibble with the numeric variables centered and/or scaled
}
\description{
This function takes a tibble \code{x} and scales the numeric variables. By default, the variables are centered and scaled to have unit variance. The function returns a tibble with the same attributes as the input tibble.
}
\examples{
df_scale(mtcars)

}
