% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download RedCap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the API token stored in .REnviron.}

\item{redcapUrl}{The URL of the RedCap instance.}

\item{redcapReportId}{The ID of the RedCap report to query.}
}
\value{
A tibble containing the contents of the RedCap report.
}
\description{
This function queries a RedCap report using the provided API token, URL, and report ID,
and returns the contents as a tibble.
}
\examples{
# Example usage
redcap_report <- downloadRedcapReport("redcap_token", "https://redcap.emory.edu/api/", "46524")
}
