% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculate F1 score of a binary classification model}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A binary vector of predicted labels}

\item{truth}{A binary vector of true labels}
}
\value{
The F1 score of the binary classification model
}
\description{
This function calculates the F1 score of a binary classification model based on the predicted labels \code{pred} and the true labels \code{truth}.
}
\examples{
f1(c(1,1,0,1), c(1,0,0,1))

}
