% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{Check if values in a vector are equal to the minimum value}
\usage{
is_min(x, na.rm = TRUE)
}
\arguments{
\item{x}{An atomic vector to check}

\item{na.rm}{Logical. Should missing values be removed? (default = TRUE)}
}
\value{
A logical vector with \code{TRUE} values where \code{x} equals its minimum value
}
\description{
This function accepts an atomic vector \code{x} and returns a logical vector with \code{TRUE} values where \code{x} equals its minimum value.
}
\examples{
is_min(c(3, 5, 1, 3, 2))
is_min(c(3, 5, NA, 3, 2))

}
