% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Calculate the log-likelihood of a sample under the chi-squared distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector of data}

\item{df}{The degrees of freedom of the chi-squared distribution}
}
\value{
The log-likelihood of \code{x} under the chi-squared distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under the chi-squared distribution, with degrees of freedom \code{df}.
}
\examples{
log_likelihood_chisq(rchisq(100, 5), 5)

}
