% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Calculate the log-likelihood of a sample under the F distribution}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of data}

\item{df1}{The degrees of freedom of the numerator of the F distribution}

\item{df2}{The degrees of freedom of the denominator of the F distribution}
}
\value{
The log-likelihood of \code{x} under the F distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under the F distribution, with degrees of freedom \code{df1} and \code{df2}.
}
\examples{
log_likelihood_f(rf(100, 5, 10), 5, 10)

}
