% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Calculate the log-likelihood of a sample under the normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector of data}

\item{mean}{The mean of the normal distribution}

\item{sd}{The standard deviation of the normal distribution}
}
\value{
The log-likelihood of \code{x} under the normal distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under the normal distribution, with mean \code{mean} and standard deviation \code{sd}.
}
\examples{
log_likelihood_norm(rnorm(100), 0, 1)

}
