% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculate the log-likelihood of a sample under the t distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector of data}

\item{df}{The degrees of freedom of the t distribution}
}
\value{
The log-likelihood of \code{x} under the t distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under the t distribution, with degrees of freedom
}
\examples{
log_likelihood_t(rt(100, 5), 5)

}
