% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Calculate the log-likelihood of a sample under the uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector of data}

\item{min}{The minimum value of the uniform distribution}

\item{max}{The maximum value of the uniform distribution}
}
\value{
The log-likelihood of \code{x} under the uniform distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under the uniform distribution, with minimum value \code{min} and maximum value \code{max}.
}
\examples{
log_likelihood_unif(runif(100), 0, 1)

}
