% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate Maximum Likelihood Estimate of p for Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of binary data (0s and 1s).}
}
\value{
The maximum likelihood estimate of \code{p}.
}
\description{
This function takes a vector of binary data and calculates the parameter \code{p}
that maximizes the log-likelihood function for a Bernoulli distribution.
}
\examples{
data <- c(1, 0, 0, 0, 1, 1, 1)
logLikBernoulli(data)

}
