% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate the minimum n per group needed for a two-sample t-test}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{The expected Cohen's d}

\item{power}{The desired statistical power (default 0.8)}
}
\value{
The minimum sample size per group needed for a two-sample t-test
}
\description{
This function calculates the minimum sample size per group needed for a two-sample t-test, given the expected effect size \code{d} and the desired statistical power \code{power}.
}
\examples{
minimum_n_per_group(0.5)

}
