% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for T-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector representing the first sample.}

\item{x2}{Optional. A numeric vector representing the second sample. If omitted, only x1 is used.}
}
\value{
The minimum sample size needed for a t-test.
}
\description{
This function acts as a wrapper around either pwr::pwr.t.test or pwr::pwr.t2n.test to calculate the minimum sample size
needed for a t-test with the given preliminary data, ensuring 80% power at alpha=0.05.
In the two sample case, it will return the minimum sample size for the second sample, but it should be
representative of the sample size required for 80% power at alpha=0.05 for both samples in a 2 sample t-test.
}
\examples{
# One sample case
x1 <- rnorm(90, mean = -3, sd = 10)
minimumN(x1)

# Two sample case
x2 <- rnorm(50, mean = 3, sd = 10)
minimumN(x1, x2)

}
