% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Randomly sample elements or rows from a vector or data frame}
\usage{
rando(x, n = 1, replace = TRUE)
}
\arguments{
\item{x}{A vector or data frame to sample from}

\item{n}{The number of samples or rows to select (default = 1)}

\item{replace}{Logical. Should sampling be done with replacement? (default = TRUE)}
}
\value{
A vector or data frame containing the sampled elements or rows
}
\description{
This function is a wrapper around the \code{sample} function that can handle both atomic vectors and data frames as input. It randomly samples either \code{n} elements from an atomic vector or \code{n} rows from a data frame, with or without replacement.
}
\examples{
rando(1:10, n = 5)
rando(mtcars, n = 3)

}
