% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate sensitivity of a binary classification model}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{A binary vector of predicted labels}

\item{truth}{A binary vector of true labels}
}
\value{
The sensitivity of the binary classification model
}
\description{
This function calculates the sensitivity of a binary classification model based on the predicted labels \code{pred} and the true labels \code{truth}.
}
\examples{
sensitivity(c(1,1,0,1), c(1,0,0,1))

}
