% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble with variable names to be standardized.}
}
\value{
A tibble with standardized variable names.
}
\description{
This function takes a tibble \code{data} and standardizes the variable names to "small_camel" case
using \code{dplyr::rename_with} and \code{janitor::make_clean_names}.
}
\examples{
data <- tibble::tibble(`First Name` = c("Alice", "Bob", "Charlie"),
               `Last Name` = c("Smith", "Johnson", "Doe"))
standardized_data <- standardizeNames(data)

}
