% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and Plot Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numerical vector representing survival status (1 for event occurred, 0 for censored).}

\item{time}{A numerical vector representing survival time.}
}
\value{
NULL. The function produces a plot of the survival curve.
}
\description{
This function takes two numerical vectors \code{status} and \code{time} representing
survival status and survival time, respectively. It calculates and plots the survival curve S(t).
}
\examples{
# Load data
data <- read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
# Extract status and time vectors from the dataset
status <- data$status
time <- data$time
# Plot survival curve
survCurv(status, time)

}
