% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse Scaling and Centering}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numerical vector that has been previously scaled.}
}
\value{
The original unscaled vector.
}
\description{
This function takes a vector that has been previously scaled (e.g., using the \code{scale()} function)
and reverses the centering and scaling.
}
\examples{
scaled_vector <- scale(1:10)
unscaled_vector <- unscale(scaled_vector)

}
