% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsup.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download Redcap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the Redcap API token.}

\item{redcapUrl}{The URL of the Redcap server.}

\item{redcapReportId}{The ID of the Redcap report to download.}
}
\value{
A tibble containing the downloaded Redcap report data.
}
\description{
Using the provided API token, URL, and report ID, this function queries a Redcap
server to download a specific report and returns the contents as a tibble.
}
