% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsup.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate Maximum Likelihood Estimate (MLE) for Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of binary outcomes (0s and 1s).}
}
\value{
The parameter p that maximizes the log-likelihood.
}
\description{
logLikBernoulli function calculates the parameter p that maximizes the log-likelihood for a Bernoulli distribution.
}
\examples{
data <- c(1, 0, 0, 0, 1, 1, 1)
logLikBernoulli(data)
}
