% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsup.R
\name{minimumN}
\alias{minimumN}
\title{Minimum Sample Size Calculation for T-Test}
\usage{
minimumN(x1, x2 = NULL, power = 0.8, sig_level = 0.05)
}
\arguments{
\item{x1}{Numeric vector representing the first sample.}

\item{x2}{Numeric vector representing the second sample (optional).}

\item{power}{Desired statistical power (default is 0.8).}

\item{sig_level}{Significance level (default is 0.05).}
}
\value{
Minimum sample size needed for the t-test.
}
\description{
This function calculates the minimum sample size needed for a t-test with power 80% and significance level 0.05 (alpha).
}
\examples{
# Calculate minimum sample size for one-sample t-test
min_n_1 <- minimumN(x1 = rnorm(30))
min_n_1
# Calculate minimum sample size for two-sample t-test
min_n_2 <- minimumN(x1 = rnorm(30), x2 = rnorm(30))
min_n_2
}
