% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsup.R
\name{survCurv}
\alias{survCurv}
\title{survCurv function calculates and plots a survival curve S(t) based on the provided status and time vectors.}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numerical vector representing censoring status (0 for censored, 1 for event).}

\item{time}{A numerical vector representing survival times.}
}
\value{
Returns Kaplan-Meir survival curve plot.
}
\description{
survCurv function calculates and plots a survival curve S(t) based on the provided status and time vectors.
}
\examples{
status <- c(1, 1, 0, 1, 0, 1, 0)
time <- c(10, 15, 20, 25, 30, 35, 40)
survCurv(status, time)
}
