% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsup.R
\name{unscale}
\alias{unscale}
\title{Reverse Centering and Scaling}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled using 'scale' function.}
}
\value{
The unscaled vector.
}
\description{
This function takes a vector that has been scaled and reverses the centering/scaling.
}
\examples{
scaled_vec <- scale(c(1, 2, 3, 4, 5))
unscaled_vec <- unscale(scaled_vec)
unscaled_vec
}
